requirements.txt
torch==2.0.1
transformers==4.32.0
tokenizers==0.13.3
peft==0.5.0
hydra-core==1.2.0
datasets==2.14.4
deepspeed==0.10.2
accelerate==0.22.0